feat: creating request refactor: code in modules
